Porting tests to tycho 0.25.0
Previous poms don't work any more with tycho 0.25.0 as surefire
dependencies cannot depend on things in the reactor build any more,
so I am trying to port them to 0.25. With this commit the tests
seem to run fine, but the test runtime may not be exactly the same
as before.

See also:
http://stackoverflow.com/questions/25476595/tycho-cannot-resolve-dependency-configured-in-tycho-surefire-plugin
https://wiki.eclipse.org/Tycho/Packaging_Types#eclipse-test-plugin

Signed-off-by: Felix Dorner <felix.dorner@gmail.com>